scripts/ai-team/requirements.txt
google-generativeai>=0.8.0
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
rich>=13.0.0